requirements.txt
undetected-chromedriver
pyautogui
setuptools
openai
flask-cors
flask
streamlit
python-dotenv
